+++ src/onefetch/deps/package_parser.rs
@@ -1,26 +1,26 @@
 use crate::onefetch::error::*;
 use {regex::Regex, toml::Value};
 
-pub fn cargo(contents: &str) -> Result<i32> {
+pub fn cargo(contents: &str) -> Result<usize> {
     let parsed = contents.parse::<Value>()?;
 
-    Ok(parsed["dependencies"].as_table().unwrap().len() as i32)
+    Ok(parsed["dependencies"].as_table().unwrap().len())
 }
 
-pub fn go_modules(contents: &str) -> Result<i32> {
+pub fn go_modules(contents: &str) -> Result<usize> {
     let count = Regex::new(r"v[0-9]+")?.find_iter(contents).count();
 
-    Ok(count as i32)
+    Ok(count)
 }
 
-pub fn npm(contents: &str) -> Result<i32> {
+pub fn npm(contents: &str) -> Result<usize> {
     let parsed = json::parse(contents)?;
 
-    Ok(parsed["dependencies"].len() as i32)
+    Ok(parsed["dependencies"].len())
 }
 
-pub fn pip(contents: &str) -> Result<i32> {
+pub fn pip(contents: &str) -> Result<usize> {
     let count = Regex::new(r"(^[A-z]+)|(\n[A-z]+)")?.find_iter(contents).count();
 
-    Ok(count as i32)
+    Ok(count)
 }
